+++ app/src/main/res/values/colors.xml
@@ -2,6 +2,8 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="colorPrimary">#FFC107</color>
+    <color name="colorPrimary">#4B4848</color>
     <color name="colorPrimaryVariant">#3F51B5</color>
+    <color name="startButtonColor">#707177</color>
+
 </resources>
