+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Unquote</string>
+    <string name="app_name">Trivia M</string>
 </resources>
